+++ src/instructlab/utils.py
@@ -612,7 +612,7 @@ def is_oci_repo(repo_url: str) -> bool:
     Checks if a provided repository follows the OCI registry URL syntax
     """
 
-    # TO DO: flesh this out and make it a more robust check
+    # TODO: flesh this out and make it a more robust check
     oci_url_prefix = "docker://"
     return repo_url.startswith(oci_url_prefix)
 
@@ -622,3 +622,22 @@ def is_huggingface_repo(repo_name: str) -> bool:
     # repo name should be of the format <owner>/<model>
     pattern = r"^[\w.-]+\/[\w.-]+$"
     return re.match(pattern, repo_name) is not None
+
+
+def _load_json(file_path: Path):
+    try:
+        with open(
+            file_path,
+            encoding="UTF-8",
+        ) as f:
+            return json.load(f)
+    except FileNotFoundError as e:
+        raise ValueError(f"file not found: {file_path}") from e
+    except json.JSONDecodeError as e:
+        raise ValueError(f"could not read JSON file: {file_path}") from e
+    except Exception as e:
+        raise ValueError("unexpected error occurred:") from e
+
+
+def _extract_SHA(SHAstr: str):
+    return re.search("sha256:(.*)", SHAstr)